buggy, add refined cube, need debug

+++ src/sim/softbody/SoftBody.h
@@ -91,4 +91,4 @@ tMatrix3d CalcPK1(const tMatrix3d &F);
 tMatrix3d CalcPK1_part1(const tMatrix3d &F);
 tMatrix3d CalcPK1_part2(const tMatrix3d &F);
 void GetSelectionMatrix(tMatrixXd &Sd, tMatrixXd &Sb);
-extern double gMu, gLambda;
+extern float gMu, gLambda;
